Section 16.0 of rfc2328 (OSPF) specifies that the short-path calculation to a node should be constructed with the sum of all path costs (metrics) to the node (pretty simple huh).  There is a usage of metric typified by the "max-metric router-lsa" command in many networking stacks that allows a router to gracefully "remove" itself from a topology by advertising the maximum value of metric in it's router LSAs (16 bits of "1").  In this case, the router will continue to forward any traffic sent to it while these "max-metric" LSAs are propagated through the network; at which point, the router can be taken out of service.
The correct handling of this in ospfd would use this metric as part of
the calculation, disuading other routers from using it for transit
traffic (assuming a better path exits).  Unfortunately, the ospfd
behavior is to remove these links from the SPF calculation.  This
patch changes the behavior to omit this exception handling.

Signed-off-by: JR Rivers <jrrivers@cumulusnetworks.com>