This patch adds support for allowing BGP to create and bring up neighbor sessions dynamically. The operator configures a range of neighbor addresses to which peering is allowed. The ranges are configured as subnets and multiple ranges are allowed. Each range is associated with a peer-group so that additional parameters can be configured.
BGP neighbor sessions are dynamically created when connections are initiated
by remote neighbors whose addresses fall within a configured range. The
sessions are deleted when the BGP connection terminates.

A limit on the number of neighbors allowed from each range of addresses
can be specified.

IPv4 and IPv6 peering is supported. Over the peering, any of the address
families configured for the peer-group can be negotiated.