If on-shutdown is configured to a large value and 'service quagga restart' is executed, then the init.d/quagga script doesnt wait more than 120 seconds for the daemon do stop, worse, it goes ahead and starts the new daemon regardless. This can result into two ospfd processes running on the same config. Which leads to many issues including but not limited to high cpu usage. Thats because the two processes are mixing packets on adjancencies thus causing churn on the box and network.
As long as OSPF is able to reliably send the max-metric router-lsa before
exiting thats mostly good enough for this purpose anyways.

As a solution to this situation, bringing the maximum configurable value of
the on-shutdown timer below the maximum retry to stop a daemon in init.d/quagga

Notes: This may not be an upstreamable patch, still we needed to find
a solution for init.d/quagga and this command this co-exist.