vtysh-add-mark-cmd.patch
VTYSH: Add support for marking a file with appropriate end of context

To support applying only differences to the existing config, this patch
enables supplying the appropriate end markers to a provided file (or
stdin). By end markers, I mean, adding "end" and "exit-address-family"
at the appropriate places in the configuration to ease finding the
differences with the running configuration.