Fix dynamic sessions with multiple bgp instances
Ticket:CM-6534
Reviewed by:CCR-3239
Testing: See Bug

Dynamic neighbors did not work properly with multiple bgp instances.
This was caused by the lookup for the dynamic range attempting to
break out of two for loops at the same time.  The break only stopped
the innermost for loop, causing it to only ever find the listen-range
for the last instance in the bm->bgp list.

Additional bug fix for dynamic listen for peer groups that are configured
with internal or external key word.